clean(client): redux refactor
