--- strife-client/src/actions/types.js
@@ -1,9 +1,0 @@
-// For recipients
-export const CHANGE_RECIPIENT = 'CHANGE_RECIPIENT';
-
-// For unseen messages
-export const ADD_UNSEEN = 'ADD_UNSEEN';
-export const REMOVE_UNSEEN = 'REMOVE_UNSEEN';
-
-// For call data
-export const CHANGE_CALL_DATA = 'CHANGE_CALL_DATA';
