+++ strife-client/src/reducers/root-reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import recipientReducer from './recipient-reducer.js';
 import notificationReducer from './notification-reducer.js';
-import callDataReducer from './call-data-reducer.js';
+import callDataReducer from './calldata-reducer.js';
 
 export default combineReducers({
   recipient: recipientReducer,
