+++ strife-client/src/types/calldata-types.js
@@ -1,0 +1,1 @@
+export const CHANGE_CALL_DATA = 'CHANGE_CALL_DATA';
