+++ strife-client/src/types/notification-types.js
@@ -1,0 +1,2 @@
+export const ADD_UNSEEN = 'ADD_UNSEEN';
+export const REMOVE_UNSEEN = 'REMOVE_UNSEEN';
