+++ strife-client/src/types/recipient-types.js
@@ -1,0 +1,1 @@
+export const CHANGE_RECIPIENT = 'CHANGE_RECIPIENT';
